Refactor BackFlow Slater functions into dedicated source
